requirements.txt
numpy>=1.25
Pillow>=10.0
matplotlib>=3.8